move funcs[1]
