+++ tests/unit/sync/sync/test_many_files.py
@@ -3,7 +3,7 @@ import shutil
 from pathlib import Path
 from unittest.mock import AsyncMock
 
-from dbx.commands.sync.sync import create_path_matcher
+from dbx.commands.sync.functions import create_path_matcher
 from dbx.sync import RemoteSyncer
 from tests.unit.sync.utils import temporary_directory
