+++ tests/unit/sync/sync/test_unmatched_deletes.py
@@ -4,7 +4,7 @@ from unittest.mock import AsyncMock, patch
 
 import pytest
 
-from dbx.commands.sync.sync import create_path_matcher
+from dbx.commands.sync.functions import create_path_matcher
 from dbx.sync import RemoteSyncer, DeleteUnmatchedOption
 from tests.unit.sync.utils import temporary_directory
